Organized Files and Directories
- Added MainCharacter.java, utilities, and Wands to the Worlds directory.
