--- Shop/Wand.java
@@ -1,15 +1,0 @@
-/********************************************************
-@author(s)          *insert here
-@date               *insert here
-@teacher            Andrew Carreiro
-@file               Shop.java
-@description        *insert class description here
-********************************************************/
-package Shop;
-
-public class Wand
-{
-
-    
-
-}
