+++ Worlds/SideQuests/FireQuest1.java
@@ -6,7 +6,7 @@
 @description        First side quest in the fire world.
 ********************************************************/
 
-package Worlds.SideQuests;
+package SideQuests;
 
 public class FireQuest1 extends SideQuests {
