+++ Worlds/SideQuests/WaterQuest2.java
@@ -6,7 +6,7 @@
 @description        Second side quest in the water world.
 ********************************************************/
 
-package Worlds.SideQuests;
+package SideQuests;
 
 public class WaterQuest2 extends SideQuests {
