+++ Worlds/Wand/Ebony.java
