+++ Worlds/Wand/Maple.java
