+++ Worlds/Wand/Pine.java
