+++ Worlds/Wand/Spruce.java
