+++ Worlds/Wand/Wand.java
